Solve knip errors
