+++ services/121-service/src/payments/fsp-integration/intersolve-visa/dto/intersolve-create-wallet-response-data.dto.ts
@@ -1,0 +1,5 @@
+import { IntersolveCreateWalletResponseTokenDto } from '@121-service/src/payments/fsp-integration/intersolve-visa/dto/intersolve-create-wallet-response-token.dto';
+
+export class IntersolveCreateWalletResponseDataDto {
+  public token: IntersolveCreateWalletResponseTokenDto;
+}
